Added Doctor APIs for user

+++ src/main/java/com/med/check/db/dto/response/DoctorReviewResponse.java
@@ -13,5 +13,5 @@ public class DoctorReviewResponse {
     private String name;
     private String surName;
     private String position;
-    private int grade;
+    private double grade;
 }
